+++ polars/polars-algo/Cargo.toml
@@ -11,8 +11,8 @@ description = "Algorithms built upon Polars primitives"
 
 [dependencies]
 polars-core = { version = "0.25.1", path = "../polars-core", features = ["private", "dtype-categorical", "asof_join"], default-features = false }
-polars-ops = { version = "0.25.1", path = "../polars-ops", features = ["dtype-categorical", "asof_join"], default-features = false }
 polars-lazy = { version = "0.25.1", path = "../polars-lazy", features = ["asof_join", "concat_str", "strings"] }
+polars-ops = { version = "0.25.1", path = "../polars-ops", features = ["dtype-categorical", "asof_join"], default-features = false }
 
 [package.metadata.docs.rs]
 all-features = true
