+++ polars/polars-core/src/named_from.rs
@@ -465,9 +465,9 @@ mod test {
         // check if we can construct.
         let _df = df![
             "date" => [NaiveDate::from_ymd(2021, 1, 1)],
-            "datetime" => [NaiveDate::from_ymd(2021, 1, 1).and_hms(0, 0, 0)],
+            "datetime" => [NaiveDate::from_ymd_opt(2021, 1, 1).unwrap().and_hms_opt(0, 0, 0).unwrap()],
             "optional_date" => [Some(NaiveDate::from_ymd(2021, 1, 1))],
-            "optional_datetime" => [Some(NaiveDate::from_ymd(2021, 1, 1).and_hms(0, 0, 0))],
+            "optional_datetime" => [Some(NaiveDate::from_ymd_opt(2021, 1, 1).unwrap().and_hms_opt(0, 0, 0).unwrap())],
             "time" => [NaiveTime::from_hms_opt(23, 23, 23).unwrap()],
             "optional_time" => [Some(NaiveTime::from_hms_opt(23, 23, 23).unwrap())],
             "duration" => [ChronoDuration::from_std(std::time::Duration::from_secs(10)).unwrap()],
