+++ polars/tests/it/core/mod.rs
@@ -1,3 +1,4 @@
+mod date_like;
 mod groupby;
 mod joins;
 mod list;
